#2 - Creating a neo4j module to pre-format requests to the neo4j database.
